delete api

+++ src/example/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render
 from django.http import JsonResponse
+from django.shortcuts import get_object_or_404
 
 from rest_framework import status
 from rest_framework.views import APIView
@@ -31,13 +32,15 @@ class ListPersonView(APIView):
                 'message': 'Create a new person unsuccessfully'
             }, status = status.HTTP_400_BAD_REQUEST)
         
-     def delete(self, request, format=None, id=None):
-        try:
-            person = Person.objects.get(pk=id)
-                person.delete()
-            return Response({'massage' : 'massage deleted'})
-        except exceptions.ObjectDoesNotExist:
-            return Response({'message':'Object does not exist'}, status=status.HTTP_204_NO_CONTENT)
+    def delete(self, request, *args, **kwargs):
+        person = get_object_or_404(Person, id=kwargs.get('pk'))
+
+        if person is not None:
+            person.delete()
+
+            return JsonResponse({
+                'message': 'delete 1 person successfully'
+            }, status = status.HTTP_200_OK)
